Added Iframe demo to the spec
All requirements for this project have been completed per the spec. Please  see the iframe added to the bottom of the original readme.md which is the spec for the project. In the project files you will also see the readme2.md which is the readme.md file created as a result of the project meeting all requirements.